Current iteration of chatroom pre frontend
This version of the python based chatroom is pre frontend. It contains a server that actively tracks incoming and outgoing users, a client that allows for usernames and messages to be communicated with other users. and a means to save all messages and pull them via a chat_data.json file.